+++ packages/data-schema/src/runtime/internals/operations/get.ts
@@ -119,9 +119,11 @@ async function _get(
     // flatten response
     if (data) {
       const [key] = Object.keys(data);
-      const flattenedResult = flattenItems(data)[key];
+      const flattenedResult = flattenItems(modelIntrospection, name, data[key]);
 
-      if (options?.selectionSet) {
+      if (flattenedResult === null) {
+        return { data: null, extensions };
+      } else if (options?.selectionSet) {
         return { data: flattenedResult, extensions };
       } else {
         // TODO: refactor to avoid destructuring here
@@ -155,7 +157,7 @@ async function _get(
      */
     if (data && Object.keys(data).length !== 0 && errors) {
       const [key] = Object.keys(data);
-      const flattenedResult = flattenItems(data)[key];
+      const flattenedResult = flattenItems(modelIntrospection, name, data[key]);
 
       /**
        * `flattenedResult` could be `null` here (e.g. `data: { getPost: null }`)